Update portable PDB indexing scheme and fix formatting
This updates the proposed indexing scheme for portable PDBs to use the MVID plus UInt32.MaxValue as the age. This allows clients that are currently downloading Windows PDBs using symsrv.dll to pretty trivially download portable PDBs.

This also fixes some formatting problems with the .md files in this repo since GitHub now requires a space between the `#` character and the title.