Remove named .gnu.debug.build-id section assumption
The code to read build-ids from ELF files assumes the build-id is in a
section named .gnu.debug.build-id. This doesn't necessarily have to be
true. In many firmware projects, the build-id needs to be placed
somewhere in particular and so a different section name is used.

Also, the SSQP protocol document doesn't defined a restriction on the
section name either. It just states:

> ELF Note section that is named “GNU” and that has note type
> PRPSINFO (3)

Which is still validated