Inspect ELF to determine if symbol file
When generating the Key for an ELF, the ELF-buildid-sym format

> applies to any ELF format files that have not been stripped of
> debugging information

This means that any ELF file, regardless of it's extension, should
generate a ELF-buildid-sym key if it has debugging symbols in it.

This changes the logic to detect if symbols exist or not by looking for
the .debug_info section.