Pointer layout changes
Delay getting a Pointer's target type layout until the pointer's target is dereferenced to avoid recursion/stack overflow
when creating the layout for types that have a field that a Pointer to itself.

This does change some protected fields in the PointerLayout (_storageLayout, _targetLayout) and Pointer (_targetLayout) classes
but I think the chances are very low that some one depends on them.

Added LayoutManager.AddReflectionTypes override with a typeFilter that allows duplicated code to be removed from the DCR.